backend/requirements-voice.txt
Flask==3.1.1
requests==2.32.4
openai-whisper==20231117
torch==2.1.0
numpy>=1.20.0
tqdm>=4.66.0
transformers>=4.32.0
pyaudio>=0.2.11
pydub>=0.25.1
SpeechRecognition>=3.10.0
